modules/guru-intelligence/src/collectors/google_trends_collector.py
"""Google Trends data collector"""
import logging
from typing import List, Dict
from datetime import datetime, timedelta
from pytrends.request import TrendReq
import asyncio

logger = logging.getLogger(__name__)

class GoogleTrendsCollector:
    """Collects trending SEO topics from Google Trends"""
    
    def __init__(self):
        self.pytrends = TrendReq(hl='en-US', tz=360)
    
    async def collect(self) -> List[Dict]:
        """Collect trending topics"""
        try:
            logger.info("Collecting Google Trends data...")
            
            # Get trending searches
            trending = self.pytrends.trending_searches(pn='united_states')
            
            results = []
            for trend in trending.head(10)[0]:  # Top 10 trends
                results.append({
                    'keyword': trend,
                    'source': 'google_trends',
                    'collected_at': datetime.utcnow().isoformat(),
                    'relevance_score': 0.8
                })
            
            logger.info(f"Collected {len(results)} trends from Google")
            return results
            
        except Exception as e:
            logger.error(f"Google Trends collection failed: {e}")
            return []
